Status_mailer/requirements.txt
pandas
configparser
apscheduler
snowflake-connector-python
sqlalchemy
cryptography